+++ src/main/java/it/codedvalue/springboot/campaign/repository/CampaignRepository.java
@@ -13,6 +13,6 @@ import org.springframework.data.repository.reactive.ReactiveCrudRepository;
 import org.springframework.stereotype.Repository;
 import reactor.core.publisher.Flux;
 
-public interface CampaignRepository extends CrudRepository<Campaign, Long> {
+public interface CampaignRepository extends CrudRepository<Campaign, Integer> {
 
 }
